move cli-related code (not working)
