+++ src/bin/whadido.ts
@@ -1,0 +1,3 @@
+import { main } from '../cli';
+
+main();
