+++ src/error.ts
@@ -1,0 +1,4 @@
+/**
+ * Error when handling git repo
+ */
+export class GitRepoException extends Error {}
